Implemented a far less excessive CPU operation handling system, instead of having 50% more completely empty Actions than actions with actual code in them, it just skips over empty CPU cycles.
Too bad TickBuses is currently running like 8 million memory reads/writes per second and kills performance

+++ FemBoyLib/Memory/MemoryDMG.cs
@@ -26,7 +26,7 @@ public class DotMatrixRAM : IMemory {
             
             // Cart RAM
             case (>=0xA000 and <= 0xBFFF):
-                return gameboy.Cartridge.Read((ushort)(address));
+                return gameboy.Cartridge.Read(address);
             
             // WRAM
             case (>=0xC000 and <= 0xDFFF):
@@ -69,7 +69,7 @@ public class DotMatrixRAM : IMemory {
         
             // Cart RAM
             case (>= 0xA000 and <= 0xBFFF):
-                gameboy.Cartridge.Write((ushort)(address), value); 
+                gameboy.Cartridge.Write(address, value); 
                 break;
         
             // WRAM
